requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
schedule==1.2.1
